refactor ticket components and remove qr deps
