+++ src/components/admin/TicketTemplateSettings.jsx
@@ -2,7 +2,7 @@ import React, { useState, useRef, useEffect, useMemo } from 'react';
 import * as FiIcons from 'react-icons/fi';
 import SafeIcon from '../../common/SafeIcon';
 import supabase from '../../lib/supabase';
-import { downloadTicketsPNG } from '../../utils/ticketExport.js';
+import { downloadTicketsPNG } from '../../utils/ticketExport';
 import { formatDateTime } from '../../utils/formatDateTime';
 import TicketLayoutSettings from './TicketLayoutSettings';
 import SMTPSettings from './SMTPSettings';
@@ -110,9 +110,6 @@ const TicketTemplateSettings = () => {
         : lastSoldTicket.zone
           ? 'zone'
           : 'general',
-      qrValue: lastSoldTicket.id
-        ? `T-${lastSoldTicket.id.substring(0, 8)}`
-        : '',
     };
   }, [lastSoldTicket, templateSettings]);
