+++ src/components/ticket/index.js
@@ -1,2 +1,2 @@
-export { default as TicketTemplate } from './TicketTemplate.jsx';
-export { default as TicketDesigner } from './TicketDesigner.jsx';
+export { default as TicketTemplate } from './TicketTemplate';
+export { default as TicketDesigner } from './TicketDesigner';
